move gradient to utils
